api_rest/requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cors==1.0.1
deezer-python==5.3.3
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==21.12.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.48.1
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tldextract==3.3.0
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
